requirements.txt
aws-cdk-lib==2.149.0
constructs>=10.0.0,<11.0.0
gradio==4.40.0
boto3==1.34.140
pillow==10.4.0
fastapi==0.111.0
mangum==0.17.0
pandas==2.2.2
